Move `SerializeSelf` into the `encoding` module
