+++ lttcore/src/encoding.rs
@@ -1,15 +1,52 @@
-//! Serialize/Deserialize strategies
+//! Serialize/Deserialize support
 //!
 //! This module includes the [`Encoding`] enum which can be used to serialize/deserialize
 //! messages. Tooling working with storage/transmission of `LibTableTop` things should do so
-//! through this module
+//! through this module.
 
 use bytes::Bytes;
 use serde::{de::DeserializeOwned, Deserialize, Serialize};
 use std::fmt::Debug;
 use std::hash::Hash;
 
-/// A {de}serialization format for wire/disk uses
+/// Turn `&self` to [`Bytes`]
+///
+/// This trait exists so it can be used behind a trait object. This has the pretty serious tradeoff
+/// that we lost type information need to deserialize the output.
+pub trait SerializeSelf {
+    /// Turn `self` into [`Bytes`]
+    ///
+    /// ```
+    /// use lttcore::encoding::{SerializeSelf, Encoding};
+    ///
+    /// assert_eq!(
+    ///   [1,2,3].serialize_self(Encoding::Json).expect("could encode"),
+    ///   "[1,2,3]"
+    /// );
+    /// ```
+    fn serialize_self(&self, encoding: Encoding) -> Result<Bytes, EncodingError>;
+
+    /// Helper function to **approximate** the type that produced the output
+    ///
+    /// This defaults to [`std::any::type_name`] which shouldn't be relied upon, it's also the best
+    /// we have 🤷
+    ///
+    /// We can get bytes via the [`SerializeSelf::serialize_self`] method, but it's really
+    /// difficult to go backwards to the original struct implementing [`SerializeSelf`] when we're
+    /// using the trait through `dyn` (most of the time). This is a hack to provide slightly better
+    /// error messages, output comes with the same tradeoffs as [`std::any::type_name`]
+    fn source_hint(&self) -> &'static str {
+        std::any::type_name::<Self>()
+    }
+}
+
+impl<T: Serialize> SerializeSelf for T {
+    fn serialize_self(&self, encoding: Encoding) -> Result<Bytes, EncodingError> {
+        encoding.serialize(self)
+    }
+}
+
+/// A specifier for which {de}serialization format the wire/disk uses
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
 pub enum Encoding {
     /// [Bincode encoding](https://docs.rs/bincode/1.3.3/bincode/)
